refactor: account recovery to be usable on verification
